Auto-format code

+++ controller/helpers/RecordingHelper.py
@@ -4,10 +4,11 @@ Created on Tue Apr 14 15:17:52 2020
 
 @author: Testa4
 """
-from pyqtgraph.Qt import QtCore
-import numpy as np
-import h5py
 import time
+
+import h5py
+import numpy as np
+from pyqtgraph.Qt import QtCore
 
 from controller.enums import RecMode
 
@@ -16,20 +17,20 @@ class RecordingHelper():
     def __init__(self, comm_channel, cameraHelper):
         self.__comm_channel = comm_channel
         self.__cameraHelper = cameraHelper
-        self.__record = False 
-    
+        self.__record = False
+
     @property
     def record(self):
         return self.__record
-        
+
     @property
     def cameraHelper(self):
         return self.__cameraHelper
-        
+
     @property
     def comm_channel(self):
         return self.__comm_channel
-            
+
     def startRecording(self, recMode, savename, attrs, frames=None, time=None):
         self.__record = True
         self.__recordingWorker = RecordingWorker(self)
@@ -43,23 +44,24 @@ class RecordingHelper():
         self.__recordingWorker.time = time
         self.__cameraHelper.updateCameraIndices()
         self.__thread.start()
-        
+
     def endRecording(self):
         self.__record = False
         self.__thread.quit()
         self.__thread.wait()
-                   
+
     def snap(self, savename, attrs):
-        store_file = h5py.File(savename+'.hdf5', 'w', track_order=True)
+        store_file = h5py.File(savename + '.hdf5', 'w', track_order=True)
         for key in attrs.keys():
             store_file.attrs[key] = attrs[key]
         size = self.__cameraHelper.shapes
-        d = store_file.create_dataset('data', (size[0], size[1]),  dtype = 'i2')
+        d = store_file.create_dataset('data', (size[0], size[1]), dtype='i2')
         umxpx = attrs['Camera_pixel_size']
-        d.attrs["element_size_um"] = [1, umxpx, umxpx] 
+        d.attrs["element_size_um"] = [1, umxpx, umxpx]
         d[:, :] = self.__cameraHelper.image
         store_file.close()
-        
+
+
 class RecordingWorker(QtCore.QObject):
     def __init__(self, recordingHelper, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -69,27 +71,28 @@ class RecordingWorker(QtCore.QObject):
         self.running = True
         it = 0
         size = self.__recordingHelper.cameraHelper.shapes
-       
-        self.f = h5py.File(self.savename+'.hdf5', 'w')
-        d = self.f.create_dataset('data', (1, size[0], size[1]), maxshape=(None, size[0], size[1]), dtype='i2')
+
+        self.f = h5py.File(self.savename + '.hdf5', 'w')
+        d = self.f.create_dataset('data', (1, size[0], size[1]), maxshape=(None, size[0], size[1]),
+                                  dtype='i2')
         for key in self.attrs.keys():
             self.f.attrs[key] = self.attrs[key]
         umxpx = self.attrs['Camera_pixel_size']
         d.attrs["element_size_um"] = [1, umxpx, umxpx]
-        
+
         if self.recMode == RecMode.SpecFrames:
             frames = self.frames
-            while it<frames:
+            while it < frames:
                 newframes = self.__recordingHelper.cameraHelper.getChunk()
                 n = len(newframes)
                 if n > 0:
-                    if (it+n)<=frames:
-                        d.resize(n + it, axis = 0)
-                        d[it:it+n, :,:] = np.array(newframes)
-                        it += n 
+                    if (it + n) <= frames:
+                        d.resize(n + it, axis=0)
+                        d[it:it + n, :, :] = np.array(newframes)
+                        it += n
                     else:
-                        d.resize(frames, axis = 0)     
-                        d[it:frames, :, :] = np.array(newframes[0:frames-it])
+                        d.resize(frames, axis=0)
+                        d[it:frames, :, :] = np.array(newframes[0:frames - it])
                         it = frames
                     self.__recordingHelper.comm_channel.updateRecFrameNumber(it)
             self.__recordingHelper.comm_channel.updateRecFrameNumber(0)
@@ -102,23 +105,23 @@ class RecordingWorker(QtCore.QObject):
                 newframes = self.__recordingHelper.cameraHelper.getChunk()
                 n = len(newframes)
                 if n > 0:
-                    d.resize(n + it, axis = 0)
-                    d[it:it+n, :,:] = np.array(newframes)
-                    it += n 
-                    self.__recordingHelper.comm_channel.updateRecTime(np.around(current, decimals=2))
+                    d.resize(n + it, axis=0)
+                    d[it:it + n, :, :] = np.array(newframes)
+                    it += n
+                    self.__recordingHelper.comm_channel.updateRecTime(
+                        np.around(current, decimals=2)
+                    )
                     current = time.time() - start
             self.__recordingHelper.comm_channel.updateRecTime(0)
             self.__recordingHelper.comm_channel.endRecording()
             self.__recordingHelper.endRecording()
-                
+
         else:
             while self.__recordingHelper.record:
                 newframes = self.__recordingHelper.cameraHelper.getChunk()
-                n = len(newframes) 
+                n = len(newframes)
                 if n > 0:
-                    d.resize(n + it, axis = 0)
-                    d[it:it+n, :,:] = np.array(newframes)
-                    it += n   
+                    d.resize(n + it, axis=0)
+                    d[it:it + n, :, :] = np.array(newframes)
+                    it += n
             self.f.close()
-
-
